requirements.txt
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
